+++ RELEASES.md
@@ -1,4 +1,4 @@
-# Release v0.5.6b 2026-02-11
+# Release v0.5.6b 2026-02-12
 
 ## New Features
 
@@ -8,6 +8,9 @@
   `CONFIG SET server_binding` or `server_port`
 4. add `ASYNCH` option to `START` function to restart server on the fly without causing deadlock errors 
    - downside is that server may not be started when function exits (v0.5.6b 2026-02-11)
+
+## Fixes
+1. Make `STOP` asynch only when called from a remote interface that runs through the internal server (v0.5.6b 2026-02-12)
 
 # Release v0.5.5b 2026-02-06